app/bot/requirements.txt
# Core library for the Telegram Bot (versión estable recomendada)
python-telegram-bot==21.1.1

# For handling SSH connections to devices
paramiko==3.4.0

# Dependency for Paramiko, bueno tenerla fijada
cryptography==42.0.8

# For robust handling of time zones (añadido para consistencia de fechas)
pytz

# For loading .env file during local development
python-dotenv==1.0.1
